# charging-docks-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Electric vehicle Charging Station API",
        "description": "Electric vehicle Charging Station API",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1"
        }
    ],
    "paths": {
        "/company": {
            "get": {
                "tags": [
                    "list"
                ],
                "summary": "Retrieve all companies",
                "description": "Get list of companies",
                "operationId": "listCompany",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "create"
                ],
                "summary": "Add new company",
                "description": "Add new company",
                "operationId": "d7bd964d23e0ec8f55cf3b4e73948627",
                "requestBody": {
                    "description": "Pass company details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "parent company name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/company/{id}": {
            "get": {
                "tags": [
                    "find by id"
                ],
                "summary": "Get company information",
                "description": "Returns company data",
                "operationId": "getCompanyById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "company id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "update by id"
                ],
                "summary": "Edit company",
                "description": "Update company",
                "operationId": "UpdateById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "company id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Update company details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "parent_company_name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "parent company name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Delete by id"
                ],
                "summary": "Delete company",
                "description": "Delete company",
                "operationId": "DeleteById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "company id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "API Endpoints",
            "description": "API Endpoints of Projects"
        }
    ]
}